requirements.txt
amadeus==12.0.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
cryptography==44.0.2
Django==5.1.7
django-allauth==65.7.0
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
fire==0.7.0
greenlet==3.1.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
phonenumbers==9.0.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyYAML==6.0.2
rav==0.0.9
requests==2.32.3
rich==13.9.4
sqlparse==0.5.3
stripe==12.0.1
termcolor==2.5.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
